refactor serialization => serialization/json
